refactor ebpf interceptor.
