+++ pkg/cni/cli/config.go
@@ -1,4 +1,4 @@
-package plugin
+package cli
 
 import (
 	"encoding/json"
