+++ pkg/cni/cli/kube.go
@@ -1,4 +1,4 @@
-package plugin
+package cli
 
 import (
 	"context"
