+++ pkg/cni/cli/log.go
@@ -1,4 +1,4 @@
-package plugin
+package cli
 
 import "os"
