+++ pkg/cni/cli/types.go
@@ -1,5 +1,5 @@
 // Package plugin implements ecnet cni plugin.
-package plugin
+package cli
 
 import (
 	"github.com/flomesh-io/fsm/pkg/logger"
