+++ pkg/cni/server/controller/handler.go
@@ -1,4 +1,4 @@
-package cniserver
+package controller
 
 import (
 	"encoding/json"
