+++ pkg/cni/server/controller/install.go
@@ -1,4 +1,4 @@
-package cniserver
+package controller
 
 import (
 	"bytes"
