+++ pkg/cni/server/controller/server.go
@@ -1,4 +1,4 @@
-package cniserver
+package controller
 
 import (
 	"context"
